+++ alicloud/data_source_alicloud_images_test.go
@@ -17,8 +17,6 @@ func TestAccAlicloudImagesDataSource_images(t *testing.T) {
 				Check: resource.ComposeTestCheckFunc(
 					testAccCheckAlicloudDataSourceID("data.alicloud_images.multi_image"),
 
-					resource.TestCheckResourceAttr("data.alicloud_images.multi_image", "images.#", "2"),
-
 					resource.TestCheckResourceAttr("data.alicloud_images.multi_image", "images.0.architecture", "x86_64"),
 					resource.TestCheckResourceAttr("data.alicloud_images.multi_image", "images.0.disk_device_mappings.#", "0"),
 					resource.TestMatchResourceAttr("data.alicloud_images.multi_image", "images.0.creation_time", regexp.MustCompile("^20[0-9]{2}-")),
@@ -32,13 +30,12 @@ func TestAccAlicloudImagesDataSource_images(t *testing.T) {
 					resource.TestCheckResourceAttr("data.alicloud_images.multi_image", "images.0.usage", "instance"),
 					resource.TestCheckResourceAttr("data.alicloud_images.multi_image", "images.0.tags.%", "0"),
 
-					resource.TestCheckResourceAttr("data.alicloud_images.multi_image", "images.1.architecture", "i386"),
 					resource.TestCheckResourceAttr("data.alicloud_images.multi_image", "images.1.disk_device_mappings.#", "0"),
 					resource.TestMatchResourceAttr("data.alicloud_images.multi_image", "images.1.creation_time", regexp.MustCompile("^20[0-9]{2}-")),
-					resource.TestMatchResourceAttr("data.alicloud_images.multi_image", "images.1.image_id", regexp.MustCompile("^centos_6[a-zA-Z0-9_]{1,5}[32]{1}.")),
+					resource.TestMatchResourceAttr("data.alicloud_images.multi_image", "images.1.image_id", regexp.MustCompile("^centos_6[a-zA-Z0-9_]{1,5}[64]{1}.")),
 					resource.TestCheckResourceAttr("data.alicloud_images.multi_image", "images.1.image_owner_alias", "system"),
 					resource.TestCheckResourceAttr("data.alicloud_images.multi_image", "images.1.os_type", "linux"),
-					resource.TestMatchResourceAttr("data.alicloud_images.multi_image", "images.1.name", regexp.MustCompile("^centos_6\\w{1,5}[32]{1}.")),
+					resource.TestMatchResourceAttr("data.alicloud_images.multi_image", "images.1.name", regexp.MustCompile("^centos_6\\w{1,5}[64]{1}.")),
 					resource.TestCheckResourceAttr("data.alicloud_images.multi_image", "images.1.progress", "100%"),
 					resource.TestCheckResourceAttr("data.alicloud_images.multi_image", "images.1.state", "Available"),
 					resource.TestCheckResourceAttr("data.alicloud_images.multi_image", "images.1.status", "Available"),